+++ app/controllers/projects/field_slips_controller.rb
@@ -11,29 +11,38 @@ module Projects
 
     def new
       flash_error(:field_slip_no_project.t) unless find_project!
-      @page_max = page_max
+      @field_slip_max = field_slip_max
     end
 
     def create
       return unless find_project!
-      return unless ok_page_count
+      return unless ok_field_slip_count
 
-      tracker = FieldSlipJobTracker.create(prefix: @project.field_slip_prefix,
-                                           start: @project.next_field_slip,
-                                           title: @project.title,
-                                           user: User.current,
-                                           count: 6 * pages)
+      tracker = FieldSlipJobTracker.create(
+        prefix: @project.field_slip_prefix,
+        start: @project.next_field_slip,
+        title: @project.title,
+        user: User.current,
+        one_per_page: params[:one_per_page] == "1",
+        count: field_slips
+      )
       if tracker
-        @project.next_field_slip = tracker.last + 1
-        if @project.save
-          FieldSlipJob.perform_later(tracker.id)
-          flash_notice(:field_slips_created_job.t(filename: tracker.filename))
-        else
-          tracker.destroy
-          flash_error(:field_slips_project_update_fail.t(title: @project.title))
-        end
+        update_project(tracker)
       else
         flash_error(:field_slips_tracker_fail.t(title: @project.title))
+      end
+    end
+
+    private
+
+    def update_project(tracker)
+      @project.next_field_slip = tracker.last + 1
+      if @project.save
+        FieldSlipJob.perform_later(tracker.id)
+        flash_notice(:field_slips_created_job.t(filename: tracker.filename))
+      else
+        tracker.destroy
+        flash_error(:field_slips_project_update_fail.t(title: @project.title))
       end
       # redirect_to(project_url(@project))
       respond_to do |format|
@@ -49,33 +58,31 @@ module Projects
       end
     end
 
-    private
-
     def find_project!
       @project = find_or_goto_index(Project, params[:project_id].to_s)
     end
 
-    def pages
-      @pages ||= params[:pages].to_i
+    def field_slips
+      @field_slips ||= params[:field_slips].to_i
     end
 
-    def page_max
-      @page_max ||= if @project.is_admin?(User.current)
-                      2000
-                    elsif @project.member?(User.current)
-                      10
-                    else
-                      0
-                    end
+    def field_slip_max
+      @field_slip_max ||= if @project.is_admin?(User.current)
+                            12_000
+                          elsif @project.member?(User.current)
+                            60
+                          else
+                            0
+                          end
     end
 
-    def ok_page_count
-      return true if page_max.positive? && pages <= page_max
+    def ok_field_slip_count
+      return true if field_slip_max.positive? && field_slips <= field_slip_max
 
-      if page_max.zero?
+      if field_slip_max.zero?
         flash_error(:field_slips_must_be_member.t)
       else
-        flash_error(:field_slips_too_many_pages.t(max: page_max))
+        flash_error(:field_slips_too_many_field_slips.t(max: field_slip_max))
       end
       redirect_to(new_project_field_slip_path(project_id: @project.id))
       false
