+++ db/migrate/20240504122906_add_one_per_page_to_field_slip_job_tracker.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+class AddOnePerPageToFieldSlipJobTracker < ActiveRecord::Migration[7.1]
+  def change
+    add_column(:field_slip_job_trackers, :one_per_page, :boolean,
+               null: false, default: false)
+  end
+end
