+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[7.1].define(version: 2024_04_23_204346) do
+ActiveRecord::Schema[7.1].define(version: 2024_05_04_122906) do
   create_table "api_keys", id: :integer, charset: "utf8mb3", force: :cascade do |t|
     t.datetime "created_at", precision: nil
     t.datetime "last_used", precision: nil
@@ -95,6 +95,7 @@ ActiveRecord::Schema[7.1].define(version: 2024_04_23_204346) do
     t.integer "pages", default: 0, null: false
     t.string "title", limit: 100, default: "", null: false
     t.integer "user_id"
+    t.boolean "one_per_page", default: false, null: false
   end
 
   create_table "field_slips", charset: "utf8mb4", collation: "utf8mb4_0900_ai_ci", force: :cascade do |t|
